Add ai to chatbot, sua ui

+++ src/main/java/com/shop/clothingstore/dto/ProductFilterDTO.java
@@ -1,12 +1,15 @@
 package com.shop.clothingstore.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class ProductFilterDTO {
+
     private Long categoryId;
     private Long subCategoryId;
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private String keyword;
 }
