+++ src/main/java/com/shop/clothingstore/dto/api/OrderResponse.java
@@ -1,5 +1,6 @@
 package com.shop.clothingstore.dto.api;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -9,6 +10,10 @@ import com.shop.clothingstore.entity.Order;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+/**
+ * DTO trả về thông tin đơn hàng qua REST API. FIXED: Dùng BigDecimal cho total
+ * và price — tránh floating-point imprecision khi serialize JSON.
+ */
 @Data
 @AllArgsConstructor
 public class OrderResponse {
@@ -18,7 +23,10 @@ public class OrderResponse {
     private String phone;
     private String address;
     private String status;
-    private Double total;
+
+    // FIXED: BigDecimal thay Double — tài chính không được dùng Double
+    private BigDecimal total;
+
     private List<OrderItemInfo> items;
 
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
@@ -31,10 +39,17 @@ public class OrderResponse {
         private String productName;
         private String size;
         private String color;
-        private Double price;
+
+        // FIXED: BigDecimal thay Double
+        private BigDecimal price;
+
         private Integer quantity;
     }
 
+    /**
+     * Map Order entity → OrderResponse DTO. Không expose entity trực tiếp ra
+     * ngoài API.
+     */
     public static OrderResponse from(Order order) {
 
         List<OrderItemInfo> items = order.getItems() != null
@@ -43,7 +58,7 @@ public class OrderResponse {
                         i.getProductName(),
                         i.getSize(),
                         i.getColor(),
-                        i.getPrice(),
+                        i.getPrice(), // BigDecimal → BigDecimal, no conversion needed
                         i.getQuantity()
                 ))
                         .toList()
@@ -55,7 +70,7 @@ public class OrderResponse {
                 order.getPhone(),
                 order.getAddress(),
                 order.getStatus().name(),
-                order.getTotal(),
+                order.getTotal(), // BigDecimal → BigDecimal
                 items,
                 order.getCreatedAt()
         );
